Added Final Version of Design
Fair Notice: This design is only one approach of many, using my knowledge of Computer Science, I decided to help with a Camel Up Project coded in C++. Using the code provided and adding new classes and methods, I've created a design approach that utilizes Encapsulation, OOP, Readability and Reliability. The code is commented to dictate what each function is used for, and the method and variables are declared in simple terms to make more know of usage. The Game works as originally intended, however the game now plays until there is a winner and the camels move in the correct direction and positions are tracked and modified when needed. The Camels also correctly order themselves from the Bottom to Top judging on who landed on who, the game now includes more than one round that are composed of 5 turns. The other actions were not implemented and that is for the ugvelez to decide if the design approach is acceptable and can be understood and modified. ReadMe will not be built for the sake of time, however and explanation of the entire project can be given by myself if needed.